+++ Common/Require-AppInstallerMinimumVersion.ps1
@@ -6,7 +6,7 @@
     .NOTES
         Author:   Olav Rønnestad Birkeland
         Created:  220307
-        Modified: 220307
+        Modified: 220316
 
     .EXAMPLE
         & $psISE.CurrentFile.FullPath
@@ -27,19 +27,11 @@ $MinimumVersion = [System.Version] '1.17.10271.0'
 
 
 # Detect and exit
+$RunningAsSystem = [bool]([System.Security.Principal.WindowsIdentity]::GetCurrent().'User'.'Value' -eq 'S-1-5-18')
 $InstalledVersion = [System.Version](
     $(
         Try {
-            if ([System.Security.Principal.WindowsIdentity]::GetCurrent().'User'.'Value' -eq 'S-1-5-18') {
-                $(
-                    [string](
-                        (Get-Item -Path ('{0}\WindowsApps\Microsoft.DesktopAppInstaller_*_x64__8wekyb3d8bbwe' -f $env:ProgramW6432)).'FullName' | Select-Object -First 1
-                    )
-                ).Split('_')[-4]
-            }
-            else {
-                (Get-AppxPackage -Name 'Microsoft.DesktopAppInstaller').'Version'
-            }
+            (Get-AppxPackage -AllUsers:$RunningAsSystem -Name 'Microsoft.DesktopAppInstaller').'Version'
         }
         Catch {
             '0.0.0'